refactor and move crate
