+++ openhcl/underhill_core/Cargo.toml
@@ -58,6 +58,7 @@ hyperv_secure_boot_templates.workspace = true
 hyperv_uefi_custom_vars_json.workspace = true
 framebuffer.workspace = true
 hcl_compat_uefi_nvram_storage = { workspace = true, features = ["inspect"] }
+hcl_mapper.workspace = true
 get_helpers.workspace = true
 get_lower_vtl.workspace = true
 get_protocol.workspace = true
